step 3

+++ pkg/models/account.go
@@ -1,7 +1,7 @@
 package models
 
 type Account struct {
-	Name   string `json:"name"`
-	Cash   int    `json:"cash"`
-	UserId int    `json:"id"`
+	Name   string  `json:"name"`
+	Cash   float64 `json:"cash"`
+	UserId int     `json:"id"`
 }
